+++ database/migrations/2025_05_04_064544_create_transaksi_keluar_table.php
@@ -18,8 +18,8 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained('users')->onDelete('cascade'); // Add user_id column
             $table->date('tanggal');
             $table->integer('qty');
-            $table->decimal('harga', 15, 2);
-            $table->decimal('jumlah', 15, 2);
+            $table->integer('harga');
+            $table->integer('jumlah');
             $table->text('keterangan')->nullable();
             $table->timestamps();
